Refactor css helper to a new util file
